+++ test/JsonMachineTest/StreamChunksTest.php
@@ -5,6 +5,9 @@ namespace JsonMachineTest;
 use JsonMachine\Exception\InvalidArgumentException;
 use JsonMachine\StreamChunks;
 
+/**
+ * @covers \JsonMachine\StreamChunks
+ */
 class StreamChunksTest extends \PHPUnit_Framework_TestCase
 {
     public function testThrowsIfNoResource()