+++ test/bootstrap.php
@@ -2,6 +2,6 @@
 
 require_once __DIR__.'/../vendor/autoload.php';
 
-if (! class_exists('PHPUnit_Framework_TestCase')) {
+if ( ! class_exists('PHPUnit_Framework_TestCase')) {
     class_alias(\PHPUnit\Framework\TestCase::class, 'PHPUnit_Framework_TestCase');
 }